StreamSymTypeRelations cleanup
